+++ build.gradle.kts
@@ -94,7 +94,7 @@ tasks.register<JavaExec>("runGenerator") {
     classpath = sourceSets["main"].runtimeClasspath
 }
 
-tasks.register<JavaExec>("runJobViaServer") {
-    mainClass.set("com.wgslfuzz.tools.RunJobViaServerKt")
+tasks.register<JavaExec>("runJobsViaServer") {
+    mainClass.set("com.wgslfuzz.tools.RunJobsViaServerKt")
     classpath = sourceSets["main"].runtimeClasspath
 }
